refactor: maualInput 관련 메서드 Manual 클래스로 리팩토링
